tenderchad_scraper/requirements.txt
aspose-words==23.8.0
boto3==1.28.27
fastapi==0.100.0
httptools==0.6.0
httpx==0.24.1
pip-chill==1.0.3
python-docx==0.8.11
python-dotenv==1.0.0
python-environ==0.4.54
pyyaml==6.0
rarfile==4.0
scrapy==2.9.0
unrar==0.4
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
# annotated-types==0.5.0 # Installed as dependency for pydantic
# anyio==3.7.1 # Installed as dependency for httpcore, starlette, watchfiles
# attrs==23.1.0 # Installed as dependency for automat, service-identity, twisted
# automat==22.10.0 # Installed as dependency for twisted
# botocore==1.31.35 # Installed as dependency for boto3, s3transfer
# certifi==2023.7.22 # Installed as dependency for httpcore, httpx, requests
# cffi==1.15.1 # Installed as dependency for cryptography
# charset-normalizer==3.2.0 # Installed as dependency for requests
# click==8.1.7 # Installed as dependency for uvicorn
# constantly==15.1.0 # Installed as dependency for twisted
# cryptography==41.0.3 # Installed as dependency for pyopenssl, scrapy, service-identity
# cssselect==1.2.0 # Installed as dependency for parsel, scrapy
# exceptiongroup==1.1.3 # Installed as dependency for anyio
# filelock==3.12.2 # Installed as dependency for tldextract
# h11==0.14.0 # Installed as dependency for httpcore, uvicorn
# httpcore==0.17.3 # Installed as dependency for httpx
# hyperlink==21.0.0 # Installed as dependency for twisted
# idna==3.4 # Installed as dependency for anyio, httpx, hyperlink, requests, tldextract
# incremental==22.10.0 # Installed as dependency for twisted
# itemadapter==0.8.0 # Installed as dependency for itemloaders, scrapy
# itemloaders==1.1.0 # Installed as dependency for scrapy
# jmespath==1.0.1 # Installed as dependency for boto3, botocore, itemloaders, parsel
# lxml==4.9.3 # Installed as dependency for parsel, python-docx, scrapy
# packaging==23.1 # Installed as dependency for parsel, scrapy
# parsel==1.8.1 # Installed as dependency for itemloaders, scrapy
# protego==0.3.0 # Installed as dependency for scrapy
# pyasn1==0.5.0 # Installed as dependency for pyasn1-modules, service-identity
# pyasn1-modules==0.3.0 # Installed as dependency for service-identity
# pycparser==2.21 # Installed as dependency for cffi
# pydantic==2.3.0 # Installed as dependency for fastapi
# pydantic-core==2.6.3 # Installed as dependency for pydantic
# pydispatcher==2.0.7 # Installed as dependency for scrapy
# pyopenssl==23.2.0 # Installed as dependency for scrapy
# python-dateutil==2.8.2 # Installed as dependency for botocore
# queuelib==1.6.2 # Installed as dependency for scrapy
# requests==2.31.0 # Installed as dependency for requests-file, tldextract
# requests-file==1.5.1 # Installed as dependency for tldextract
# s3transfer==0.6.2 # Installed as dependency for boto3
# service-identity==23.1.0 # Installed as dependency for scrapy
# six==1.16.0 # Installed as dependency for automat, python-dateutil, requests-file
# sniffio==1.3.0 # Installed as dependency for anyio, httpcore, httpx
# starlette==0.27.0 # Installed as dependency for fastapi
# tldextract==3.4.4 # Installed as dependency for scrapy
# twisted==22.10.0 # Installed as dependency for scrapy
# typing-extensions==4.7.1 # Installed as dependency for fastapi, pydantic, pydantic-core, twisted
# urllib3==1.26.16 # Installed as dependency for botocore, requests
# w3lib==2.1.2 # Installed as dependency for itemloaders, parsel, scrapy
# zope.interface==6.0 # Installed as dependency for scrapy, twisted
